BREAKING CHANGE: rename top-level files to camel case
