+++ lib/browser.js
@@ -4,7 +4,7 @@
 
 require('./driver').set(require('./drivers/browser'));
 
-const DocumentProvider = require('./document_provider.js');
+const DocumentProvider = require('./documentProvider.js');
 
 DocumentProvider.setBrowser(true);
 
@@ -67,7 +67,7 @@ exports.Types = require('./types');
  * @method VirtualType
  * @api public
  */
-exports.VirtualType = require('./virtualtype');
+exports.VirtualType = require('./virtualType');
 
 /**
  * The various Mongoose SchemaTypes.
@@ -81,7 +81,7 @@ exports.VirtualType = require('./virtualtype');
  * @api public
  */
 
-exports.SchemaType = require('./schematype.js');
+exports.SchemaType = require('./schemaType.js');
 
 /**
  * Internal utils
