+++ lib/collection.js
@@ -5,7 +5,7 @@
  */
 
 const EventEmitter = require('events').EventEmitter;
-const STATES = require('./connectionstate');
+const STATES = require('./connectionState');
 const immediate = require('./helpers/immediate');
 
 /**
