+++ lib/connectionState.js
