+++ lib/cursor/QueryCursor.js
@@ -7,7 +7,7 @@
 const MongooseError = require('../error/mongooseError');
 const Readable = require('stream').Readable;
 const eachAsync = require('../helpers/cursor/eachAsync');
-const helpers = require('../queryhelpers');
+const helpers = require('../queryHelpers');
 const kareem = require('kareem');
 const immediate = require('../helpers/immediate');
 const util = require('util');
