+++ lib/document.js
@@ -34,7 +34,7 @@ const inspect = require('util').inspect;
 const internalToObjectOptions = require('./options').internalToObjectOptions;
 const markArraySubdocsPopulated = require('./helpers/populate/markArraySubdocsPopulated');
 const mpath = require('mpath');
-const queryhelpers = require('./queryhelpers');
+const queryhelpers = require('./queryHelpers');
 const utils = require('./utils');
 const isPromise = require('./helpers/isPromise');
