+++ lib/documentProvider.js
@@ -15,7 +15,7 @@ let isBrowser = false;
  *
  * @api private
  */
-module.exports = function() {
+module.exports = function documentProvider() {
   if (isBrowser) {
     return BrowserDocument;
   }
