+++ lib/drivers/node-mongodb-native/connection.js
@@ -6,7 +6,7 @@
 
 const MongooseConnection = require('../../connection');
 const MongooseError = require('../../error/index');
-const STATES = require('../../connectionstate');
+const STATES = require('../../connectionState');
 const mongodb = require('mongodb');
 const pkg = require('../../../package.json');
 const processConnectionOptions = require('../../helpers/processConnectionOptions');
