+++ lib/helpers/query/completeMany.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const helpers = require('../../queryhelpers');
+const helpers = require('../../queryHelpers');
 
 module.exports = completeMany;
