+++ lib/index.js
@@ -10,11 +10,11 @@ const Document = require('./document');
 const EventEmitter = require('events').EventEmitter;
 const Kareem = require('kareem');
 const Schema = require('./schema');
-const SchemaType = require('./schematype');
+const SchemaType = require('./schemaType');
 const SchemaTypes = require('./schema/index');
-const VirtualType = require('./virtualtype');
-const STATES = require('./connectionstate');
-const VALID_OPTIONS = require('./validoptions');
+const VirtualType = require('./virtualType');
+const STATES = require('./connectionState');
+const VALID_OPTIONS = require('./validOptions');
 const Types = require('./types');
 const Query = require('./query');
 const Model = require('./model');
@@ -981,7 +981,7 @@ Mongoose.prototype.Document = Document;
  * @api public
  */
 
-Mongoose.prototype.DocumentProvider = require('./document_provider');
+Mongoose.prototype.DocumentProvider = require('./documentProvider');
 
 /**
  * The Mongoose ObjectId [SchemaType](https://mongoosejs.com/docs/schematypes.html). Used for
