+++ lib/internal.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const StateMachine = require('./statemachine');
+const StateMachine = require('./stateMachine');
 const ActiveRoster = StateMachine.ctor('require', 'modify', 'init', 'default', 'ignore');
 
 module.exports = exports = InternalCache;
