+++ lib/model.js
@@ -62,7 +62,7 @@ const prepareDiscriminatorPipeline = require('./helpers/aggregate/prepareDiscrim
 const pushNestedArrayPaths = require('./helpers/model/pushNestedArrayPaths');
 const removeDeselectedForeignField = require('./helpers/populate/removeDeselectedForeignField');
 const setDottedPath = require('./helpers/path/setDottedPath');
-const STATES = require('./connectionstate');
+const STATES = require('./connectionState');
 const util = require('util');
 const utils = require('./utils');
 const MongooseBulkWriteError = require('./error/bulkWriteError');
@@ -3823,7 +3823,7 @@ Model.hydrate = function(obj, projection, options) {
     obj = applyProjection(obj, projection);
   }
 
-  const document = require('./queryhelpers').createModel(this, obj, projection);
+  const document = require('./queryHelpers').createModel(this, obj, projection);
   document.$init(obj, options);
   return document;
 };
