+++ lib/query.js
@@ -22,7 +22,7 @@ const clone = require('./helpers/clone');
 const completeMany = require('./helpers/query/completeMany');
 const getDiscriminatorByValue = require('./helpers/discriminator/getDiscriminatorByValue');
 const hasDollarKeys = require('./helpers/query/hasDollarKeys');
-const helpers = require('./queryhelpers');
+const helpers = require('./queryHelpers');
 const immediate = require('./helpers/immediate');
 const internalToObjectOptions = require('./options').internalToObjectOptions;
 const isExclusive = require('./helpers/projection/isExclusive');
