+++ lib/queryHelpers.js
