+++ lib/schema.js
@@ -7,10 +7,10 @@
 const EventEmitter = require('events').EventEmitter;
 const Kareem = require('kareem');
 const MongooseError = require('./error/mongooseError');
-const SchemaType = require('./schematype');
+const SchemaType = require('./schemaType');
 const SchemaTypeOptions = require('./options/SchemaTypeOptions');
 const VirtualOptions = require('./options/VirtualOptions');
-const VirtualType = require('./virtualtype');
+const VirtualType = require('./virtualType');
 const addAutoId = require('./helpers/schema/addAutoId');
 const clone = require('./helpers/clone');
 const get = require('./helpers/get');
