+++ lib/schema/DocumentArrayElement.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const MongooseError = require('../error/mongooseError');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const SubdocumentPath = require('./SubdocumentPath');
 const getConstructor = require('../helpers/discriminator/getConstructor');
