+++ lib/schema/SubdocumentPath.js
@@ -8,7 +8,7 @@ const CastError = require('../error/cast');
 const EventEmitter = require('events').EventEmitter;
 const ObjectExpectedError = require('../error/objectExpected');
 const SchemaSubdocumentOptions = require('../options/SchemaSubdocumentOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const applyDefaults = require('../helpers/document/applyDefaults');
 const $exists = require('./operators/exists');
 const castToNumber = require('./operators/helpers').castToNumber;
