+++ lib/schema/array.js
@@ -8,7 +8,7 @@ const $exists = require('./operators/exists');
 const $type = require('./operators/type');
 const MongooseError = require('../error/mongooseError');
 const SchemaArrayOptions = require('../options/SchemaArrayOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const CastError = SchemaType.CastError;
 const Mixed = require('./mixed');
 const arrayDepth = require('../helpers/arrayDepth');
