+++ lib/schema/bigint.js
@@ -5,7 +5,7 @@
  */
 
 const CastError = require('../error/cast');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const castBigInt = require('../cast/bigint');
 const utils = require('../utils');
