+++ lib/schema/boolean.js
@@ -5,7 +5,7 @@
  */
 
 const CastError = require('../error/cast');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const castBoolean = require('../cast/boolean');
 const utils = require('../utils');
