+++ lib/schema/buffer.js
@@ -6,7 +6,7 @@
 
 const MongooseBuffer = require('../types/buffer');
 const SchemaBufferOptions = require('../options/SchemaBufferOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const handleBitwiseOperator = require('./operators/bitwise');
 const utils = require('../utils');
