+++ lib/schema/date.js
@@ -6,7 +6,7 @@
 
 const MongooseError = require('../error/index');
 const SchemaDateOptions = require('../options/SchemaDateOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const castDate = require('../cast/date');
 const getConstructorName = require('../helpers/getConstructorName');
 const utils = require('../utils');
