+++ lib/schema/decimal128.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const CastError = SchemaType.CastError;
 const castDecimal128 = require('../cast/decimal128');
 const utils = require('../utils');
