+++ lib/schema/documentarray.js
@@ -10,7 +10,7 @@ const DocumentArrayElement = require('./DocumentArrayElement');
 const EventEmitter = require('events').EventEmitter;
 const SchemaDocumentArrayOptions =
   require('../options/SchemaDocumentArrayOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const discriminator = require('../helpers/model/discriminator');
 const handleIdOption = require('../helpers/schema/handleIdOption');
 const handleSpreadDoc = require('../helpers/document/handleSpreadDoc');
