+++ lib/schema/map.js
@@ -6,7 +6,7 @@
 
 const MongooseMap = require('../types/map');
 const SchemaMapOptions = require('../options/SchemaMapOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 /*!
  * ignore
  */
