+++ lib/schema/mixed.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const symbols = require('./symbols');
 const isObject = require('../helpers/isObject');
 const utils = require('../utils');
