+++ lib/schema/number.js
@@ -6,7 +6,7 @@
 
 const MongooseError = require('../error/index');
 const SchemaNumberOptions = require('../options/SchemaNumberOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const castNumber = require('../cast/number');
 const handleBitwiseOperator = require('./operators/bitwise');
 const utils = require('../utils');
