+++ lib/schema/objectid.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const SchemaObjectIdOptions = require('../options/SchemaObjectIdOptions');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const castObjectId = require('../cast/objectid');
 const getConstructorName = require('../helpers/getConstructorName');
 const oid = require('../types/objectid');
