+++ lib/schema/string.js
@@ -4,7 +4,7 @@
  * Module dependencies.
  */
 
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const MongooseError = require('../error/index');
 const SchemaStringOptions = require('../options/SchemaStringOptions');
 const castString = require('../cast/string');
