+++ lib/schema/uuid.js
@@ -5,7 +5,7 @@
 'use strict';
 
 const MongooseBuffer = require('../types/buffer');
-const SchemaType = require('../schematype');
+const SchemaType = require('../schemaType');
 const CastError = SchemaType.CastError;
 const utils = require('../utils');
 const handleBitwiseOperator = require('./operators/bitwise');
