+++ lib/schemaType.js
