+++ lib/stateMachine.js
