+++ lib/validOptions.js
