+++ lib/virtualType.js
